test(checkbox-card): recategorize and clean up browser tests
